FixKnowNew

+++ DiamondLuxurySolution/DiamondLuxurySolution.Application/Repository/Product/ProductRepo.cs
@@ -85,7 +85,7 @@ namespace DiamondLuxurySolution.Application.Repository.Product
                 {
                     if (item.effectDate.Date == DateTime.Now.Date)
                     {
-                        totalPriceGem += item.Price;
+                        totalPriceGem += (decimal)item.Price;
                     }
                 }
                 if (errorList.Any())
@@ -105,7 +105,7 @@ namespace DiamondLuxurySolution.Application.Repository.Product
                     {
                         if (item.effectDate.Date == DateTime.Now.Date)
                         {
-                            totalMaterialPrice += item.SellPrice;
+                            totalMaterialPrice += (decimal)item.SellPrice;
                         }
                     }
                     if (errorList.Any())
@@ -317,7 +317,7 @@ namespace DiamondLuxurySolution.Application.Repository.Product
                 {
                     if (item.effectDate.Date == DateTime.Now.Date)
                     {
-                        totalPriceGem += item.Price;
+                        totalPriceGem +=  (decimal)item.Price;
                     }
                 }
                 product.GemId = gem.GemId;
@@ -340,7 +340,7 @@ namespace DiamondLuxurySolution.Application.Repository.Product
                     {
                         if (item.effectDate.Date == DateTime.Now.Date)
                         {
-                            totalMaterialPrice += item.SellPrice;
+                            totalMaterialPrice += (decimal)item.SellPrice;
                         }
                     }
